requirements.txt
Django
djangorestframework
openai
chromadb
python-dotenv
langchain
pandas
requests